+++ mk_test_env/__manifest__.py
@@ -10,7 +10,7 @@
 {
     'name': 'Manage Test Environment',
     'summary': 'Create or update test environment',
-    'version': '10.0.0.6.5',
+    'version': '10.0.0.6.7',
     'category': 'Tools',
     'author': 'SHS-AV s.r.l.',
     'website': 'https://www.zeroincombenze.it/servizi-le-imprese/',